сделал разбивку на модули в стиле ES6
